1.54 Bungee/MM Hooks, sql, durability and trigger classificaiton

+++ src/com/sucy/skill/api/event/PlayerExperienceGainEvent.java
@@ -98,7 +98,7 @@ public class PlayerExperienceGainEvent extends Event implements Cancellable
      *
      * @throws IllegalArgumentException if experience is less than 0
      */
-    public void setExp(int amount)
+    public void setExp(double amount)
     {
         if (amount < 0)
         {
